isolated clightning and lnd lspd code
